test2_bot_trade/requirements.txt
sqlalchemy
pandas
numpy
requests
python-dotenv
pydantic-settings
pydantic
yfinance

tensorflow
scikit-learn
matplotlib
seaborn
ta
scipy
streamlit
plotly
transformers
torch
newsapi-python
python-telegram-bot
arch
feedparser
beautifulsoup4
nest_asyncio
psutil
pytest
pytest-cov
pytest-mock
fastapi
uvicorn
pyyaml
keyring
cryptography
